tracing: Replace `SpanKind` with a single span
In case of workflows and activities, use `worker` span to
collect the execution lifetime as well as logs.
In case of webhooks, use the topmost `incoming webhook request`
for the same purpose.
Not having a separate business span means the child span
has the most specific enabled parent, while collecting the logs
is still easy.