tracing: Work around `Span::current` giving a disabled span
With a configuration like this:

```toml
[otlp]
level = "info"

[log.stdout]
level = "info,obeli_sk_wasm_workers=trace"
```

the current span would be a debug span, even though only info spans
are enabled by the tracing layer. This caused a disconnect between
the parent and child execution.
The workaround is to attempt to serialize the metadata first using the
current span. If no metadata is injected, use an info-level span that
belongs to the current worker / webhook request.