feat!: Implement structured concurency for child executions
Workflows now wait for all child executions submitted to their
join sets before marking the parent execution as successful
or failed with a `FatalError`.
This means that intermittent errors or timeouts cannot turn
into a final execution error and must be retried util one
from the two states mentioned above are reached.
This concern is not implemented yet.