fix: Fix `null` deserialization into `WastValWithType`
When deserializing a tagged enum `JoinSetResponse`, the process
would fail with `invalid type: null`. This happens with type
`result<option<string>, string>` and value `{"ok": null}`.
The root cause is that serde calls `visit_unit` instead of `visit_none`.
It looks like this only happens when inside a tagged enum.