fix: Prioritize oci digest when tag is specified as well
Previously digest was removed and `latest` tag was added if missing.
This leads to either failing to download the image (when `latest` is not present),
or fetching the possibly wrong image and then failing on digest comparison.
New behavior is consistent with `docker pull` - when digest is specified,
the tag is disregarded.